LayoutTests/http/tests/adClickAttribution/send-attribution-conversion-request-expected.txt

Tests sending of ad click attribution requests after a conversion.



--------
Frame: '<!--frame1-->'
--------
Conversion received.
HTTP_HOST: 127.0.0.1:8000
REQUEST_URI: /adClickAttribution/resources/conversionReport.php?conversion=12&campaign=3
No cookies in conversion request.

No stored Ad Click Attribution data.
